// src/models/ComparisonOperators.ts
export enum ComparisonOperators {
    Equals = '==',
    CaseInsensitiveEquals = '==*',
    CountEquals = '#==',    
    NotEquals = '!=',
    CaseInsensitiveNotEquals = '!=*',
    CountNotEquals = '#!=',    
    GreaterThan = '>',
    CountGreaterThan = '#>',    
    LessThan = '<',
    CountLessThan = '#<',    
    GreaterThanOrEqual = '>=',
    CountGreaterThanOrEqual = '#>=',    
    LessThanOrEqual = '<=',
    CountLessThanOrEqual = '#<=',    
    StartsWith = '_=',
    CaseInsensitiveStartsWith = '_=*',    
    DoesNotStartWith = '!_=',
    CaseInsensitiveDoesNotStartWith = '!_=*',    
    EndsWith = '_-=',
    CaseInsensitiveEndsWith = '_-=*',    
    DoesNotEndWith = '!_-=',
    CaseInsensitiveDoesNotEndWith = '!_-=*',    
    Contains = '@=',
    CaseInsensitiveContains = '@=*',    
    DoesNotContain = '!@=',
    CaseInsensitiveDoesNotContain = '!@=*',    
    SoundsLike = '~~',
    DoesNotSoundLike = '!~',    
    Has = '^$',
    CaseInsensitiveHas = '^$*',    
    DoesNotHave = '!^$',
    CaseInsensitiveDoesNotHave = '!^$*',    
    In = '^^',
    CaseInsensitiveIn = '^^*',
  }